LayoutTests/platform/wpe/imported/w3c/web-platform-tests/websockets/closing-handshake/003-expected.txt


FAIL WebSockets: client sends closing handshake assert_equals: e.code expected 1005 but got 1011
